Move minimap render logic into renderLogic.ts in utils.
